Remove multiprocessing calls to printer method
The multiprocessing feature has been removed because of its inefficiency
compared to the single process counterpart. This is probably due to the
fact that the printer method is too short and there are also too much
I/O operations in order for multiprocessing to get an edge. Furthermore,
the way multiprocessing is implemented causes bugs on Windows during
file operations that crash the program raising a Permission Error.